Reorganizes functions tests.
